test(#14) Splits test_database/test_item.py into submodules
