+++ tests/__init__.py
@@ -20,9 +20,17 @@ from .test_current_user import TestCurrentUser
 from .test_migrations import TestMigrations
 from .test_file import TestFile
 from .test_rest_api_connector import TestRestApiConnector
-from .test_database import TestDatabaseEngine, TestDatabaseItem
 from .test_utils import (
     TestNestedDataPathFind,
     TestNestedDataPathUpdate,
     TestNestedDataPathDelete,
+)
+from .test_database import (
+    TestDatabaseEngine,
+    TestDatabaseItem,
+    TestDatabaseItemSearch,
+    TestDatabaseItemCreate,
+    TestDatabaseItemUpdate,
+    TestDatabaseItemDelete,
+    TestDatabaseItemLoad,
 )
