+++ tests/test_database/__init__.py
@@ -1,3 +1,15 @@
+"""All databases tests.
+
+They can be run with unittest tests.test_database
+"""
+
 from .test_engine import TestDatabaseEngine
-from .test_item import TestDatabaseItem
 from .test_attribute import TestDatabaseAttribute
+from .test_item import (
+    TestDatabaseItem,
+    TestDatabaseItemSearch,
+    TestDatabaseItemCreate,
+    TestDatabaseItemUpdate,
+    TestDatabaseItemDelete,
+    TestDatabaseItemLoad,
+)
