+++ tests/test_database/test_item/__init__.py
@@ -1,0 +1,11 @@
+"""All database item tests.
+
+They can be run with unittest tests.test_database.test_item
+"""
+
+from .test_base import TestDatabaseItem
+from .test_search import TestDatabaseItemSearch
+from .test_create import TestDatabaseItemCreate
+from .test_delete import TestDatabaseItemDelete
+from .test_update import TestDatabaseItemUpdate
+from .test_load import TestDatabaseItemLoad
